pyproject.toml
[tool.poetry]
name = "ogu_schedule"
version = "0.1.0"
description = ""
authors = ["CraazzzyyFoxx <andrey.pod19@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^2.19"
tortoise-orm = {extras = ["aiosqlite"], version = "^0.19.2"}
pydantic = "^1.10.1"
python-dotenv = "^0.20.0"
aiosqlite = "^0.17.0"
tabulate = "^0.8.10"
aiohttp = "^3.8.1"
python-json-logger = "^2.0.4"
colorlog = "^6.7.0"
selenium = "^4.4.3"
fake-useragent = "^0.1.11"
PyVirtualDisplay = "^3.0"
loguru = "^0.6.0"
click = "^8.1.3"
aiohttp_autoreload = "^0.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
